fastapi-guide/assets/production/requirements.txt
# FastAPI Production Requirements
fastapi[standard]>=0.100.0
pydantic>=2.0
pydantic-settings>=2.0
sqlmodel>=0.0.14

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.0.0
httpx>=0.24.0

# Development
uvicorn[standard]>=0.23.0
